Implement organization of `codegen`, and add style argument
