+++ src/main.rs
@@ -15,11 +15,9 @@ use upload::upload_asset;
 use crate::config::Config;
 
 pub mod args;
-mod ast;
 mod codegen;
 pub mod config;
 pub mod lockfile;
-mod nested;
 pub mod state;
 mod svg;
 mod upload;
@@ -180,7 +178,7 @@ async fn main() -> anyhow::Result<()> {
         }));
 
     let lua_filename = format!("{}.{}", state.output_name, state.lua_extension);
-    let lua_output = generate_lua(&state.new_lockfile, asset_dir_str);
+    let lua_output = generate_lua(&state.new_lockfile, asset_dir_str, &state.style);
 
     write(Path::new(&state.write_dir).join(lua_filename), lua_output?)
         .await
@@ -192,6 +190,7 @@ async fn main() -> anyhow::Result<()> {
             &state.new_lockfile,
             asset_dir_str,
             state.output_name.as_str(),
+            &state.style,
         );
 
         write(Path::new(&state.write_dir).join(ts_filename), ts_output?)
